requirements.txt
Django==2.1.2
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
bootstrap_admin==0.4.0
jsonfield==2.0.2
pyyaml== 5.1.0
django_json_widget
requests==2.20.0
six==1.11.0
websocket-client==0.53.0
jsonfield==2.0.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
